% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.metafrontier}
\alias{print.metafrontier}
\title{Print a Metafrontier Object}
\usage{
\method{print}{metafrontier}(x, ...)
}
\arguments{
\item{x}{a \code{"metafrontier"} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Metafrontier Object
}
