% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.metafrontier}
\alias{summary.metafrontier}
\title{Summary of a Metafrontier Model}
\usage{
\method{summary}{metafrontier}(object, ...)
}
\arguments{
\item{object}{a \code{"metafrontier"} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
An object of class \code{"summary.metafrontier"}.
}
\description{
Summary of a Metafrontier Model
}
